Fixed Bugs
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@11274 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/ports/SAMA/SAMA5D2x/sama_matrix.h
@@ -320,7 +320,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
- bool mtxConfigPeriphSecurity(Matrix *mtxp, uint8_t id, bool mode);
+ bool mtxConfigPeriphSecurity(Matrix *mtxp, uint32_t id, bool mode);
  void mtxConfigDefaultMaster(Matrix *mtxp, uint8_t slaveID,
                              uint8_t type, uint8_t masterID);
  void mtxConfigSlaveSec(Matrix *mtxp, uint8_t slaveID,
